server/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
torch==2.1.0
transformers==4.31.0
openai-whisper==20231117
pyttsx3==2.90
python-dotenv==0.19.0
soundfile==0.12.1
librosa==0.10.0
numpy==1.23.5